+++ core/src/main/scala/com/krux/hyperion/activity/BaseEmrActivity.scala
@@ -1,0 +1,12 @@
+package com.krux.hyperion.activity
+
+import com.krux.hyperion.resource.BaseEmrCluster
+
+/**
+ * The base trait for activities that run on an Amazon EMR cluster
+ */
+trait BaseEmrActivity[A <: BaseEmrCluster] extends PipelineActivity[A] {
+
+  type Self <: BaseEmrActivity[A]
+
+}