+++ core/src/test/scala/com/krux/hyperion/activity/SparkStepSpec.scala
@@ -19,22 +19,26 @@ class SparkStepSpec extends WordSpec {
       val ja = SparkStep(S3Uri("s3://something.jar"))
         .withMainClass(name)
         .withDriverMemory(9.gigabytes)
-      assert(ja.mainClass.map(_.toString) == Some(name))
+      assert(ja.mainClass === None)
+      assert(ja.sparkMainClass.map(_.toString) == Some(name))
     }
 
     "allow mainClass an instance" in {
       val ja = SparkStep(S3Uri("s3://something.jar")).withMainClass(new SomeClass())
-      assert(ja.mainClass.map(_.toString) == Some("com.krux.hyperion.activity.SparkStepSpec.SomeClass"))
+      assert(ja.mainClass === None)
+      assert(ja.sparkMainClass.map(_.toString) == Some("com.krux.hyperion.activity.SparkStepSpec.SomeClass"))
     }
 
     "allow mainClass an object" in {
       val ja = SparkStep(S3Uri("s3://something.jar")).withMainClass(SomeObject)
-      assert(ja.mainClass.map(_.toString) == Some("com.krux.hyperion.activity.SparkStepSpec.SomeObject"))
+      assert(ja.mainClass === None)
+      assert(ja.sparkMainClass.map(_.toString) == Some("com.krux.hyperion.activity.SparkStepSpec.SomeObject"))
     }
 
     "allow mainClass a Class" in {
       val ja = SparkStep(S3Uri("s3://something.jar")).withMainClass(SomeObject.getClass)
-      assert(ja.mainClass.map(_.toString) == Some("com.krux.hyperion.activity.SparkStepSpec.SomeObject"))
+      assert(ja.mainClass === None)
+      assert(ja.sparkMainClass.map(_.toString) == Some("com.krux.hyperion.activity.SparkStepSpec.SomeObject"))
     }
   }
 }