+++ core/src/test/scala/com/krux/hyperion/resource/EmrConfigurationSpec.scala
@@ -12,19 +12,17 @@ import org.scalatest.{FlatSpec, Matchers}
 class EmrConfigurationSpec extends FlatSpec with Matchers {
 
   it should "not emit empty lists" in {
-    implicit lazy val genEmrConfig = Arbitrary(EmrConfiguration())
+    implicit lazy val genEmrConfig = Arbitrary(EmrConfiguration("hadoop-env"))
 
     implicit lazy val genProperty = Arbitrary(for {key <- alphaStr; value <- alphaStr}
       yield Property(HType.string2HString(key), HType.string2HString(value)))
 
-    forAll { (config: List[EmrConfiguration], props: List[Property], classification: Option[String]) =>
-      val emrConfig = EmrConfiguration()
+    forAll { (config: List[EmrConfiguration], props: List[Property], classification: String) =>
+      val emrConfig = EmrConfiguration(classification)
         .withConfiguration(config: _*)
         .withProperty(props: _*)
 
-      val ser = classification
-        .fold(emrConfig)(emrConfig.withClassification(_))
-        .serialize
+      val ser = emrConfig.serialize
 
       inside(ser) { case AdpEmrConfiguration(_, _, _, prop, conf) =>
         conf shouldNot be(Some(Seq()))