+++ examples/src/main/scala/com/krux/hyperion/examples/ExampleS3DistCpWorkflow.scala
@@ -7,7 +7,7 @@ import com.typesafe.config.ConfigFactory
 import com.krux.hyperion.Implicits._
 import com.krux.hyperion.activity.S3DistCpActivity
 import com.krux.hyperion.activity.S3DistCpActivity.OutputCodec
-import com.krux.hyperion.resource.MapReduceCluster
+import com.krux.hyperion.resource.EmrCluster
 
 object ExampleS3DistCpWorkflow extends DataPipelineDef with HyperionCli {
 
@@ -31,7 +31,7 @@ object ExampleS3DistCpWorkflow extends DataPipelineDef with HyperionCli {
   override def parameters: Iterable[Parameter[_]] = Seq(target, instanceType, instanceCount, instanceBid)
 
   // Resources
-  val emrCluster = MapReduceCluster()
+  val emrCluster = EmrCluster()
     .withTaskInstanceCount(instanceCount)
     .withTaskInstanceType(instanceType)
     .withReleaseLabel("emr-4.4.0")